# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TaxonSummary {
  cdNom: number;
  cdRef: number;
  lbNom: string;
  nomVern?: string | null;
  rang?: string | null;
  famille?: string | null;
  regne?: string | null;
}

export interface TaxonDetail {
  cdNom: number;
  cdRef: number;
  cdSup?: number | null;
  regne?: string | null;
  phylum?: string | null;
  classe?: string | null;
  ordre?: string | null;
  famille?: string | null;
  group1Inpn?: string | null;
  group2Inpn?: string | null;
  group3Inpn?: string | null;
  rang?: string | null;
  lbNom: string;
  lbAuteur?: string | null;
  nomComplet?: string | null;
  nomValide?: string | null;
  nomVern?: string | null;
  nomVernEng?: string | null;
  habitat?: string | null;
  fr?: string | null;
  url?: string | null;
}

export type TaxonMediaImagesItem = {
  url: string;
  title?: string | null;
  author?: string | null;
};

export interface TaxonMedia {
  images: TaxonMediaImagesItem[];
}

export type TaxonStatsKingdomCountsItem = {
  regne: string;
  count: number;
};

export interface TaxonStats {
  totalTaxons: number;
  totalSpecies: number;
  totalGenera: number;
  totalFamilies: number;
  totalStatuts: number;
  kingdomCounts: TaxonStatsKingdomCountsItem[];
}

export interface BdcStatut {
  cdTypeStatut: string;
  lbTypeStatut?: string | null;
  regroupementType?: string | null;
  codeStatut?: string | null;
  labelStatut?: string | null;
  rqStatut?: string | null;
  cdSig?: string | null;
  lbAdmTr?: string | null;
  niveauAdmin?: string | null;
  fullCitation?: string | null;
  docUrl?: string | null;
}

export interface WikipediaInfo {
  extract?: string | null;
  url?: string | null;
  title?: string | null;
}

export interface GbifInfo {
  gbifKey?: number | null;
  occurrenceCount?: number | null;
  iucnCategory?: string | null;
  iucnCategoryLabel?: string | null;
  gbifUrl?: string | null;
  distributionCountries?: string[] | null;
}

export interface BhlReference {
  itemId?: number | null;
  titleId?: number | null;
  title: string;
  authors?: string | null;
  date?: string | null;
  url: string;
  bhlType?: string | null;
}

export interface BhlInfo {
  scientificName?: string | null;
  references?: BhlReference[];
  source?: string | null;
  fetchedAt?: string | null;
  unavailable?: boolean | null;
  message?: string | null;
}

export interface Error {
  error: string;
}

export interface BlockError {
  error: boolean;
  message: string;
}

export interface SensitivityDriver {
  label: string;
  code?: string | null;
  title: string;
  kind: string;
}

export interface SensitivityResult {
  score: number;
  label: string;
  ecological: number;
  regulatory: number;
  territorial: number;
  management: number;
  drivers: SensitivityDriver[];
}

export type ChildrenSummaryPreviewItem = {
  cdNom: number;
  lbNom: string;
  nomVern?: string | null;
  rang?: string | null;
};

export interface ChildrenSummary {
  total: number;
  preview: ChildrenSummaryPreviewItem[];
}

export interface TraitsSummary {
  hasStaticTraits: boolean;
  staticSourcesCount: number;
  staticSources: string[];
}

export type InteractionsSummaryGroupsItem = {
  id: string;
  label: string;
  count: number;
};

export interface InteractionsSummary {
  totalPartners: number;
  groups: InteractionsSummaryGroupsItem[];
}

export interface ShareSummary {
  title: string;
  description: string;
  imageUrl?: string | null;
  canonicalUrl: string;
}

export interface TaxonProfile {
  taxon: TaxonDetail;
  classification: TaxonSummary[];
  childrenSummary: ChildrenSummary;
  media: TaxonMedia | BlockError;
  statuts: BdcStatut[];
  sensitivity: SensitivityResult;
  wikipedia: WikipediaInfo | BlockError;
  gbif: GbifInfo | BlockError;
  traitsSummary: TraitsSummary;
  interactionsSummary?: InteractionsSummary | BlockError | null;
  shareSummary: ShareSummary;
  generatedAt: string;
}

export type ListStatusTypes200Item = {
  code: string;
  label: string;
  taxa: number;
};

export type ListTerritoires200Item = {
  lb: string;
  cd_sig: string;
  niveau: string;
  taxa: number;
};

export type SearchTaxonsParams = {
/**
 * Search query (name or vernacular name)
 */
q: string;
/**
 * Filter by kingdom
 */
regne?: string;
/**
 * Filter by territory cd_sig (e.g. INSEER52 region or INSEED69 département)
 */
territoire?: string;
/**
 * Max results
 */
limit?: number;
};