# lib/api-zod/src/generated/types/blockError.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockError {
  error: boolean;
  message: string;
}