# lib/api-zod/src/generated/types/childrenSummary.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */
import type { ChildrenSummaryPreviewItem } from './childrenSummaryPreviewItem';

export interface ChildrenSummary {
  total: number;
  preview: ChildrenSummaryPreviewItem[];
}