# lib/api-zod/src/generated/types/interactionsSummary.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */
import type { InteractionsSummaryGroupsItem } from './interactionsSummaryGroupsItem';

export interface InteractionsSummary {
  totalPartners: number;
  groups: InteractionsSummaryGroupsItem[];
}