# lib/api-zod/src/generated/types/sensitivityDriver.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */

export interface SensitivityDriver {
  label: string;
  code?: string | null;
  title: string;
  kind: string;
}