# lib/api-zod/src/generated/types/sensitivityResult.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */
import type { SensitivityDriver } from './sensitivityDriver';

export interface SensitivityResult {
  score: number;
  label: string;
  ecological: number;
  regulatory: number;
  territorial: number;
  management: number;
  drivers: SensitivityDriver[];
}