# lib/api-zod/src/generated/types/traitsSummary.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * TAXREF taxonomy browser API
 * OpenAPI spec version: 0.1.0
 */

export interface TraitsSummary {
  hasStaticTraits: boolean;
  staticSourcesCount: number;
  staticSources: string[];
}